* Backlog page - view story detail inline - Pulls from JSON data source * Backlog page - task grid visible inside of story - Pulls from JSON data source * Backlog page - delete story from data table (does not include the JSON data source) * Backlog page - show stories now has an 'all' option * Beginnings of the plan for Product, Release and Sprint beginning to form (not listing out @  the moment)
Needs some refactoring, I'll do that over the next two commits. And I will add in Unit Tests.

UPDATE ON FUTURE

I am aiming to complete the GUI work over the next 3 weeks, in parallel I will also be working on the DAL / Application layers, most notably the SQLite part of the project.

I have decided that the MVP (Minimal Viable Product) will consist of the following;

* Operational Task Wall
* Operational Backlog
* Operational Planning - Product, Release and Sprint
* Searching of Stories / Tasks

All of the above will be driven by the SQLite database and if possible will be throwing everything around as JSON objects.

I'm a little ahead of schedule, so I may just get the MVP out of the door for the summer.